+++ DK/DKFramework/DKAudioStream.h
@@ -2,7 +2,7 @@
 //  File: DKAudioStream.h
 //  Author: Hongtae Kim (tiff2766@gmail.com)
 //
-//  Copyright (c) 2004-2014 Hongtae Kim. All rights reserved.
+//  Copyright (c) 2004-2015 Hongtae Kim. All rights reserved.
 //
 
 #pragma once
@@ -22,7 +22,7 @@ namespace DKFramework
 	class DKGL_API DKAudioStream
 	{
 	public:
-		typedef long long Position;
+		typedef int64_t Position;
 		enum FileType
 		{
 			FileTypeUnknown = 0,
